unify SaveAnywhere's save utilities
